refactor(main): share the pack import dialog helper across modals
