+++ src/main/theme-pack-ipc.ts
@@ -5,10 +5,10 @@
 // and the main process treats pack JSON as opaque blobs that round-trip
 // via the dialog.
 
-import { BrowserWindow, dialog } from 'electron'
-import { readFile } from 'node:fs/promises'
+import { BrowserWindow } from 'electron'
 import { IpcChannels } from '../shared/ipc/channels'
 import { secureHandle } from './ipc-guard'
+import { readSelectedImportFiles } from './pack-import-dialog'
 import {
   listMetas,
   getPack,
@@ -25,29 +25,11 @@ import type {
   ThemePackRecord,
   ThemePackStoreResult,
   ThemePackImportDialogResult,
-  ThemePackImportFile,
 } from '../shared/types/theme-store'
 
 function broadcastChanged(): void {
   for (const win of BrowserWindow.getAllWindows()) {
     win.webContents.send(IpcChannels.THEME_PACK_CHANGED)
-  }
-}
-
-/** Read + parse a single file selected via the multi-select import dialog.
- *  Never throws — a read or parse failure is reported via `parseError` so
- *  one bad file in the batch does not abort the rest. */
-async function readOneImportFile(filePath: string): Promise<ThemePackImportFile> {
-  try {
-    const raw = await readFile(filePath, 'utf-8')
-    try {
-      const parsed: unknown = JSON.parse(raw)
-      return { filePath, raw: parsed, fileSizeBytes: Buffer.byteLength(raw, 'utf-8') }
-    } catch (err) {
-      return { filePath, fileSizeBytes: Buffer.byteLength(raw, 'utf-8'), parseError: String(err) }
-    }
-  } catch (err) {
-    return { filePath, parseError: String(err) }
   }
 }
 
@@ -164,22 +146,14 @@ export function setupThemePackStore(): void {
     IpcChannels.THEME_PACK_IMPORT,
     async (event): Promise<ThemePackImportDialogResult> => {
       const win = BrowserWindow.fromWebContents(event.sender)
-      if (!win) return { canceled: true, files: [] }
-      const result = await dialog.showOpenDialog(win, {
+      const files = await readSelectedImportFiles(win, {
         title: 'Import Theme Pack',
         filters: [
           { name: 'JSON', extensions: ['json'] },
           { name: 'All Files', extensions: ['*'] },
         ],
-        properties: ['openFile', 'multiSelections'],
       })
-      if (result.canceled || result.filePaths.length === 0) {
-        return { canceled: true, files: [] }
-      }
-      const files: ThemePackImportFile[] = []
-      for (const filePath of result.filePaths) {
-        files.push(await readOneImportFile(filePath))
-      }
+      if (!files) return { canceled: true, files: [] }
       return { canceled: false, files }
     },
   )
