+++ src/shared/types/theme-store.ts
@@ -93,22 +93,13 @@ export interface ThemePackStoreResult<T> {
   error?: string
 }
 
-/** One file selected via the multi-select THEME_PACK_IMPORT dialog. Parsing
- *  happens per-file in main; `parseError` is set instead of `raw` when the
- *  file could not be read or was not valid JSON. */
-export interface ThemePackImportFile {
-  filePath: string
-  raw?: unknown
-  fileSizeBytes?: number
-  parseError?: string
-}
-
-/** Shape returned by the THEME_PACK_IMPORT IPC handler. `files` is empty
- *  when `canceled` is true. */
-export interface ThemePackImportDialogResult {
-  canceled: boolean
-  files: ThemePackImportFile[]
-}
+/** Structurally identical to Language Packs' `I18nPackImportFile` /
+ *  `I18nPackImportDialogResult` — both alias the shared shape in
+ *  `pack-import.ts` (see `readSelectedImportFiles` in
+ *  `src/main/pack-import-dialog.ts`, which both IPC handlers call).
+ *  Aliased (not just re-exported) under the domain-specific name so
+ *  existing call sites don't need to change what they import. */
+export type { PackImportFile as ThemePackImportFile, PackImportDialogResult as ThemePackImportDialogResult } from './pack-import'
 
 export interface ThemePackImportApplyOptions {
   id?: string
